More bug fixes based on testvm.  - remove remaining uint32_t holdovers.  - Add ability to do "absolute" ld/st operations on packets.  - Fix netvm_rt.c to get a uint64_t return code from netvm.  - Improve error checking in testvm.  - Various other little fixes.

+++ src/netvm_rt.c
@@ -541,7 +541,7 @@ int nvmmrt_execute(struct netvm_mrt *mrt)
 	struct pktbuf *pkb;
 	struct netvm_matchedprog *mprog;
 	int i, rv, send;
-	uint32_t rc;
+	uint64_t rc;
 	struct clist_node *cln;
 
 	abort_unless(mrt);
